and nicer

+++ src/day10.rs
@@ -49,7 +49,7 @@ pub fn part2(input: &Vec<u8>) -> u128 {
 
                 let idx = n as usize;
 
-                for &i in [1 as usize, 2, 3].iter() {
+                for &i in [1, 2, 3].iter() {
                     if idx >= i {
                         acc[idx] += acc[idx - i];
                     }
